refactor(clean code): clean code

+++ include/Client.hpp
@@ -8,13 +8,13 @@
 
 class Client {
 public:
-    Client(RingBuffer<ClientMessage>& buffer, int client_id, int message_cnt, int memory_space, AccessPattern pattern);
+    Client(RingBuffer<ClientMessage>& buffer, size_t client_id, size_t message_cnt, size_t memory_space, AccessPattern pattern);
     void run();
 
 private:
     RingBuffer<ClientMessage>& buffer_;
-    int client_id_;
-    int message_cnt_;
+    size_t client_id_;
+    size_t message_cnt_;
     MemoryAccessGenerator generator_;
 };
